fix: memindahkan dan memperbaiki filter data, menghapus fs yang tak terpakai
